# docs/catalog.openapi.yaml
openapi: 3.0.4
info:
  title: Catalog service
  version: 0.1.0
servers:
  - url: http://localhost:8080/catalog
    description: Local dev server
paths:
...
  /v1/request:
    post:
      tags:
        - Request
      summary: Post a new book request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequestForm'
      responses:
        '201':
          description: Successfully posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookRequest'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Request
      summary: Get all book requests
      responses:
        '200':
          description: Successfully posted
        '500':
          $ref: '#/components/responses/500'
...
components:
  responses:
    400:
      description: Bad response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    500:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 500
        message:
          type: string
          example: internal server error
...
    BookRequest:
      type: object
      required:
        - id
        - filename
        - title
        - author_name
        - description
        - category
        - language_code
        - tags
      properties:
        id:
          type: integer
          format: int32
          example: 1
        filename:
          type: string
          example: s37djwf690seks23/doc.pdf
        title:
          type: string
          example: Harry Potter
        author_name:
          type: string
          example: J.K. Rowling
        description:
          type: string
          example: Book about Harry Potter
        category:
          type: string
          example: Fiction
        language_code:
          type: string
          example: EN
        tags:
          type: array
          items:
            type: string
          example:
            - Fantasy
            - Children
    BookRequestForm:
      type: object
      required:
        - filename
        - title
        - author_name
        - description
        - category
        - language_code
      properties:
        filename:
          type: string
          example: req_26fe31cd-77d9-415d-881f-0bfc9ed2dcc9.pdf
        title:
          type: string
          example: Harry Potter
        author_name:
          type: string
          example: J.K. Rowling
        description:
          type: string
          example: Book about Harry Potter
        category:
          type: string
          example: Fiction
        language_code:
          type: string
          example: EN
        tags:
          type: array
          items:
            type: string
          example:
            - Fantasy
            - Children
...